chore(mobile):rename screen filename
- Updated import paths for existing screens to use camel case naming conventions.
- Introduced new screens: EditEmailScreen, FollowScreen, ForgetPasswordScreen, InvitationScreen, ListScreen, LoginScreen, ProfileScreen, RsshubFormScreen, SignUpScreen, TwoFactorAuthScreen, and SelectReadingModeScreen.
- Enhanced the onboarding experience with the addition of EditProfileScreen.
