+++ apps/mobile/src/modules/entry-content/EntryReadHistory.tsx
@@ -4,7 +4,7 @@ import { Pressable, View } from "react-native"
 import { UserAvatar } from "@/src/components/ui/avatar/UserAvatar"
 import { apiClient } from "@/src/lib/api-fetch"
 import { useNavigation } from "@/src/lib/navigation/hooks"
-import { ProfileScreen } from "@/src/screens/(modal)/profile"
+import { ProfileScreen } from "@/src/screens/(modal)/ProfileScreen"
 import { useEntry } from "@/src/store/entry/hooks"
 import { isInboxEntry } from "@/src/store/entry/utils"
 import { userActions } from "@/src/store/user/store"
