+++ apps/mobile/src/modules/screen/action.tsx
@@ -17,8 +17,8 @@ import { Dialog } from "@/src/lib/dialog"
 import { useNavigation } from "@/src/lib/navigation/hooks"
 import { proxyEnv } from "@/src/lib/proxy-env"
 import { toast } from "@/src/lib/toast"
-import { LoginScreen } from "@/src/screens/(modal)/login"
-import { ProfileScreen } from "@/src/screens/(modal)/profile"
+import { LoginScreen } from "@/src/screens/(modal)/LoginScreen"
+import { ProfileScreen } from "@/src/screens/(modal)/ProfileScreen"
 import { getFeed } from "@/src/store/feed/getter"
 import { useWhoami } from "@/src/store/user/hooks"
 import { accentColor, useColor } from "@/src/theme/colors"
