+++ apps/mobile/src/modules/settings/hooks/useTOTPModalWrapper.tsx
@@ -4,7 +4,7 @@ import Siblings from "react-native-root-siblings"
 import { getFetchErrorInfo } from "@/src/lib/error-parser"
 import { useNavigation } from "@/src/lib/navigation/hooks"
 import { toast } from "@/src/lib/toast"
-import { TwoFactorAuthScreen } from "@/src/screens/(modal)/2fa"
+import { TwoFactorAuthScreen } from "@/src/screens/(modal)/TwoFactorAuthScreen"
 import { useWhoami } from "@/src/store/user/hooks"
 
 import { OTPWindow } from "../components/OTPWindow"
