+++ apps/mobile/src/modules/settings/routes/EditProfile.tsx
@@ -23,7 +23,7 @@ import { CheckCircleCuteReIcon } from "@/src/icons/check_circle_cute_re"
 import { CloseCircleFillIcon } from "@/src/icons/close_circle_fill"
 import { useNavigation } from "@/src/lib/navigation/hooks"
 import { toast } from "@/src/lib/toast"
-import { EditEmailScreen } from "@/src/screens/(modal)/edit-email"
+import { EditEmailScreen } from "@/src/screens/(modal)/EditEmailScreen"
 import { useWhoami } from "@/src/store/user/hooks"
 import type { MeModel } from "@/src/store/user/store"
 import { userSyncService } from "@/src/store/user/store"
