+++ apps/mobile/src/modules/settings/routes/Lists.tsx
@@ -26,7 +26,7 @@ import { UserAdd2CuteFiIcon } from "@/src/icons/user_add_2_cute_fi"
 import { Wallet2CuteFiIcon } from "@/src/icons/wallet_2_cute_fi"
 import { useNavigation } from "@/src/lib/navigation/hooks"
 import type { HonoApiClient } from "@/src/morph/types"
-import { ListScreen } from "@/src/screens/(modal)/list"
+import { ListScreen } from "@/src/screens/(modal)/ListScreen"
 import { useOwnedLists, usePrefetchOwnedLists } from "@/src/store/list/hooks"
 import type { ListModel } from "@/src/store/list/store"
 import { accentColor } from "@/src/theme/colors"
