+++ apps/mobile/src/screens/(modal)/EditEmailScreen.tsx
