+++ apps/mobile/src/screens/(modal)/onboarding/SelectReadingModeScreen.tsx
